// src/api/models/Location.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Location = {
    readonly location_id: number;
    readonly created_at: string;
    readonly updated_at: string;
    address_1?: string | null;
    address_2?: string | null;
    city?: string | null;
    state?: string | null;
    zip?: string | null;
    country_concept?: number | null;
};